refactor: local 디렉토리 위치 이동
- 변경 내용: entity -> dto 디렉토리로 위치 이동
- 이유: entity는 DB의 테이블과 매칭되는 JPA 객체를 다루는 디렉토리이기 때문에, 단순히 코드 안에서 사용되는 local 디렉토리는 dto 디렉토리가 더 적합할거라고 생각
