+++ src/main/java/com/writon/admin/domain/dto/lcoal/SpecialQuestionDto.java
@@ -1,4 +1,4 @@
-package com.writon.admin.domain.entity.lcoal;
+package com.writon.admin.domain.dto.lcoal;
 
 import java.util.List;
 import lombok.AllArgsConstructor;
@@ -6,8 +6,9 @@ import lombok.Getter;
 
 @Getter
 @AllArgsConstructor
-public class SpecialQuestion {
+public class SpecialQuestionDto {
 
+  private Long keywordId;
   private String keyword;
   private List<String> questions;
