+++ src/main/java/com/writon/admin/domain/dto/response/challenge/CreateChallengeResponseDto.java
@@ -1,6 +1,6 @@
 package com.writon.admin.domain.dto.response.challenge;
 
-import com.writon.admin.domain.entity.lcoal.ChallengeResponse;
+import com.writon.admin.domain.dto.lcoal.ChallengeResponse;
 import java.util.List;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
