+++ src/main/java/com/writon/admin/domain/dto/response/challenge/QuestionsResponseDto.java
@@ -1,6 +1,6 @@
 package com.writon.admin.domain.dto.response.challenge;
 
-import com.writon.admin.domain.entity.lcoal.SpecialQuestion;
+import com.writon.admin.domain.dto.lcoal.SpecialQuestionDto;
 import java.util.List;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
