+++ src/main/java/com/writon/admin/domain/service/AuthService.java
@@ -2,13 +2,11 @@ package com.writon.admin.domain.service;
 
 import com.writon.admin.domain.dto.request.auth.LoginRequestDto;
 import com.writon.admin.domain.dto.request.auth.SignUpRequestDto;
-import com.writon.admin.domain.dto.request.auth.ReissueRequestDto;
 import com.writon.admin.domain.dto.response.auth.LoginResponseDto;
-import com.writon.admin.domain.dto.response.auth.ReissueResponseDto;
 import com.writon.admin.domain.dto.response.auth.SignUpResponseDto;
 import com.writon.admin.domain.dto.wrapper.auth.LoginResponseWrapper;
 import com.writon.admin.domain.entity.challenge.Challenge;
-import com.writon.admin.domain.entity.lcoal.ChallengeResponse;
+import com.writon.admin.domain.dto.lcoal.ChallengeResponse;
 import com.writon.admin.domain.entity.organization.AdminUser;
 import com.writon.admin.domain.entity.organization.Organization;
 import com.writon.admin.domain.repository.challenge.ChallengeRepository;
@@ -18,7 +16,6 @@ import com.writon.admin.global.config.auth.TokenDto;
 import com.writon.admin.global.config.auth.TokenProvider;
 import com.writon.admin.global.error.CustomException;
 import com.writon.admin.global.error.ErrorCode;
-import jakarta.servlet.http.HttpServletRequest;
 import java.util.Collections;
 import java.util.List;
 import java.util.Optional;
