+++ src/main/java/com/writon/admin/domain/service/ChallengeService.java
@@ -5,11 +5,11 @@ import com.writon.admin.domain.dto.request.challenge.CreateChallengeRequestDto;
 import com.writon.admin.domain.dto.request.challenge.QuestionsRequestDto;
 import com.writon.admin.domain.dto.response.challenge.ChallengeInfoResponseDto;
 import com.writon.admin.domain.dto.response.challenge.QuestionsResponseDto;
-import com.writon.admin.domain.entity.lcoal.SpecialQuestion;
-import com.writon.admin.domain.entity.lcoal.ChallengeResponse;
+import com.writon.admin.domain.dto.lcoal.SpecialQuestionDto;
+import com.writon.admin.domain.dto.lcoal.ChallengeResponse;
 import com.writon.admin.domain.dto.response.challenge.CreateChallengeResponseDto;
-import com.writon.admin.domain.entity.lcoal.Status;
-import com.writon.admin.domain.entity.lcoal.UserStatus;
+import com.writon.admin.domain.dto.lcoal.Status;
+import com.writon.admin.domain.dto.lcoal.UserStatus;
 import com.writon.admin.domain.entity.activity.UserTemplate;
 import com.writon.admin.domain.entity.challenge.Challenge;
 import com.writon.admin.domain.entity.challenge.ChallengeDay;
